изменен размер аргументов сдвига и лимита

+++ mapper.go
@@ -7,7 +7,7 @@ import (
 
 type Mapper[T any] interface {
 	Find(ctx context.Context, db *sql.DB, id string) *sql.Row
-	FindAll(ctx context.Context, db *sql.DB, limit, offset int) (*sql.Rows, error)
+	FindAll(ctx context.Context, db *sql.DB, limit, offset int64) (*sql.Rows, error)
 	FindBy(ctx context.Context, db *sql.DB, conditions string, args []any) (*sql.Rows, error)
 	ExistsBy(ctx context.Context, db *sql.DB, conditions string, args []any) (bool, error)
 	CountBy(ctx context.Context, db *sql.DB, conditions string, args []any) (int64, error)
